requirements.txt
altgraph==0.17.3
async-generator==1.10
attrs==22.1.0
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clr-loader==0.2.4
colorama==0.4.5
dash==2.6.2
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-mantine-components==0.10.2
dash-table==5.0.0
et-xmlfile==1.1.0
exceptiongroup==1.0.4
Flask==2.2.2
Flask-Compress==1.13
future==0.18.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.3
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.0
pandas-datareader==0.10.0
pefile==2022.5.30
plotly==5.10.0
proxy-tools==0.1.0
pycparser==2.21
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
PySocks==1.7.1
python-dateutil==2.8.2
pythonnet==3.0.1
pytz==2022.4
pywebview==3.7.2
pywin32-ctypes==0.2.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tenacity==8.1.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
xlrd==2.0.1
